what is this... a branding!?
I found the files for it so I might as well (use them as placeholders, cuz 1. I cannot access adobe illustrator anymore, and 2. I think the branding needs an update anyway) - tho I want this to be the hybrid of a café website and a recipe website
+ topbar navigation! for now the links are just there to make it look like a legit website, but I'll see what I can do...)
= main flex

the banana bread pics are mine :)